arbfp: Allow for use of combine constant in separate blend strings
When using the texture combine constant differently in separate
RGB and alpha blend strings, the arbfp fragment backend creates a
shader that uses two different local shader parameters, but only
assigns an actual value to one of them. This patch changes this
behaviour so that only one local parameter index is used for the
texture combine constant, which is guaranteed to be the same in
both instances.

Signed-off-by: George Matsumura <gmmatsumura01@bvsd.org>